requirements.txt
# ── Core ──
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── Database ──
sqlalchemy>=2.0.25
alembic>=1.13.0
asyncpg>=0.29.0
aiosqlite>=0.19.0

# ── Auth & Security ──
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ── Cache & Queue ──
redis>=5.0.0
celery>=5.3.0

# ── HTTP & WebSocket ──
httpx>=0.26.0
websockets>=12.0

# ── AI ──
google-generativeai>=0.3.0
openai>=1.10.0

# ── Utils ──
pillow>=10.2.0
python-dateutil>=2.8.0
jinja2>=3.1.0
aiofiles>=23.2.0

# ── Monitoring ──
prometheus-client>=0.19.0
structlog>=24.1.0

# ── Testing ──
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
